Add files via upload
Word_Guess is a game which choses a random english word from a list of words and asks player to guess the word by giving appropriate hints.
The random word can be of any length. 
Player is given 8 chances to guess the word.
As hints, player is given the length of the word and is rewarded if any alphabet in guessed word and random word is similar as well as at the same position as thaat in the random word.